Refactor hybrid scan reader tests to a separate executable (#19359)
Contributes to #19308

This PR refactors gtests related to the hybrid scan reader (experimental parquet reader) to a separate executable. 

No new code is added by this PR except trivial style and CMake changes.

Authors:
  - Muhammad Haseeb (https://github.com/mhaseeb123)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Devavret Makkar (https://github.com/devavret)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/19359
